Rename foods-response-handler to be plural and pluralize filename on import lines
