+++ lib/event-listeners/foods-event-listeners.js
@@ -1,6 +1,6 @@
 import $ from 'jquery'
 import {postNewFood, deleteFoodRecord} from '../fetch-requests/foods-fetch-requests'
-import {search, missingFoodField} from '../response-handlers/foods-response-handler'
+import {search, missingFoodField} from '../response-handlers/foods-response-handlers'
 
 export function foodListeners() {
   $('#new-food-submit-button').on("click", (event) => {
