+++ lib/fetch-requests/foods-fetch-requests.js
@@ -6,7 +6,7 @@ import {
   renderNewFood,
   clearAlerts,
   clearNewFoodForm
-} from '../response-handlers/foods-response-handler'
+} from '../response-handlers/foods-response-handlers'
 
 const url = `https://protected-basin-11627.herokuapp.com/`
